Rename show-viewshed-from-point-on-map to show-viewshed-calculated-from-geoprocessing-task
