+++ samples/show-viewshed-calculated-from-geoprocessing-task/README.md
@@ -1,13 +1,14 @@
-# Show viewshed from point on map
+# Show viewshed calculated from geoprocessing task
 
 Calculate a viewshed using a geoprocessing service, in this case showing what parts of a landscape are visible from points on mountainous terrain.
 
-![Image of show viewshed from point on map](show-viewshed-from-point-on-map.png)
+![Image of show viewshed calculated from geoprocessing task](show-viewshed-calculated-from-geoprocessing-task.png)
 
 ## Use case
 
 A viewshed is used to highlight what is visible from a given point. A viewshed could be created to show what a hiker might be able to see from a given point at the top of a mountain. Equally, a viewshed could also be created from a point representing the maximum height of a proposed wind turbine to see from what areas the turbine would be visible.
 
+Note: This analysis requires a geoprocessing service endpoint in a connected environment and processing times may vary depending on connection speeds. If fast, offline analysis is required, consider using a `ViewshedFunction` to perform a viewshed instead.
 ## How to use the sample
 
 Click the map to see all areas visible from that point within a 15km radius. Clicking on an elevated and unobstructed area will highlight a larger part of the surrounding landscape. It may take a few seconds for the task to run and send back the results.
