+++ samples/show-viewshed-calculated-from-geoprocessing-task/show-viewshed-calculated-from-geoprocessing-task.png
