+++ samples/show-viewshed-calculated-from-geoprocessing-task/src/main/res/values/strings.xml
@@ -1,0 +1,3 @@
+<resources>
+    <string name="show_viewshed_calculated_from_geoprocessing_task_app_name">Show viewshed calculated from geoprocessing task</string>
+</resources>
