--- samples/show-viewshed-from-point-on-map/src/main/res/values/strings.xml
@@ -1,3 +1,0 @@
-<resources>
-    <string name="show_viewshed_from_point_on_map_app_name">Show viewshed from point on map</string>
-</resources>
